% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-odds.R
\name{get_odds}
\alias{get_odds}
\title{Get Odds}
\usage{
get_odds(
  sport = "upcoming",
  region = "us",
  mkt = "h2h",
  date_format = "unix",
  odds_format = "decimal",
  api_key = Sys.getenv("THEODDS_API_KEY")
)
}
\arguments{
\item{sport}{The sport key obtained from calling the \code{get_sports}
function. "upcoming" is always valid (and the default), returning any live
games as well as the next 8 upcoming games across all sports}

\item{region}{Determines which bookmakers are returned. Valid regions are
"au" (Australia), "uk" (United Kingdom), "eu" (Europe) and "us" (United
States). Defaults to "us"}

\item{mkt}{Determines which odds market is returned. Valid markets are "h2h"
(head to head / moneyline), "spreads" (handicaps) and "totals" (over/under).
spreads and totals odds are not always as comprehensive as h2h, so they do
not count against the usage quota on paid plans. Defaults to "h2h"}

\item{date_format}{Determines the format of timestamps in the response.
Valid values are "unix" and "iso" (ISO 8601). Defaults to "unix"}

\item{odds_format}{Determines the format of odds in the response.
Valid values are "decimal" and "american". When set to
american, small discrepancies might exist for some bookmakers due to
rounding errors. Currently, this package does not support american odds,
but hopefully will in the near future}

\item{api_key}{The API key for The Odds API -- A key can be obtained at \href{https://the-odds-api.com/}{the Odds API's website}. By default,
\code{api_key} is set to \code{Sys.getenv("THEODDS_API_KEY")}. It is highly
recommended to configure this environment variable before using this
package, both for convenience and to avoid surfacing your key as plaintext
in a script}
}
\value{
A tibble
}
\description{
The principal function of the package, \code{get_odds} allows you to pull
odds from various different sports, books, and market types into nice, tidy
tibbles
}
\details{
These function arguments correspond to the request query parameters, from
which this help page's argument descriptions were borrowed. You can see the
original documentation \href{https://the-odds-api.com/liveapi/guides/v3/#parameters-2}{on the Odds API's website}
}
\examples{
\dontrun{epl_odds <- get_odds("soccer_epl")}
\dontrun{nfl_spreads <- get_odds("americanfootball_nfl", mkt = "spreads")}
\dontrun{nba_totals <- get_odds("basketball_nba", mkt = "totals")}
}
