% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-remaining-requests.R
\name{get_remaining_requests}
\alias{get_remaining_requests}
\title{Get the Number of Remaining Requests}
\usage{
get_remaining_requests(api_key = Sys.getenv("THEODDS_API_KEY"))
}
\arguments{
\item{api_key}{The API key for The Odds API -- A key can be obtained at \href{https://the-odds-api.com/}{the Odds API's website}. By default,
\code{api_key} is set to \code{Sys.getenv("THEODDS_API_KEY")}. It is highly
recommended to configure this environment variable before using this
package, both for convenience and to avoid surfacing your key as plaintext
in a script}
}
\value{
A tibble
}
\description{
This function returns how many requests you have made for the month, and how
many are remaining. The tiers for how many requests are alotted can be found
at \href{https://the-odds-api.com/}{the Odds API website}
}
\examples{
\dontrun{get_remaining_requests()}
}
